+++ test/unit/test_pipelines.py
@@ -28,7 +28,7 @@ def test_initialize_pipeline(mock_open_project_config, mock_update_git_repo, moc
     mock_update_git_repo.assert_called_once_with(mock_webhook)
 
 
-@ mock.patch('json.loads', return_value={'mock': 'json'})
+@mock.patch('json.loads', return_value={'mock': 'json'})
 def test_open_project_config(mock_webhook):
     with mock.patch('builtins.open', mock.mock_open()):
         config = Pipeline.open_project_config(mock_webhook)
@@ -36,20 +36,21 @@ def test_open_project_config(mock_webhook):
         assert config == {'mock': 'json'}
 
 
-@ mock.patch('harvey.utils.Utils.kill')
+@mock.patch('harvey.utils.Utils.kill')
 def test_open_project_config_not_found(mock_utils_kill, mock_webhook):
     with mock.patch('builtins.open', mock.mock_open()) as mock_open:
         mock_open.side_effect = FileNotFoundError
         Pipeline.open_project_config(mock_webhook)
 
         mock_utils_kill.assert_called_once_with(
-            f'Error: Harvey could not find a "harvey.json" file in {Global.repo_full_name(mock_webhook)}.',
-            mock_webhook
+            f'Error: Harvey could not find a "harvey.json" file in {Global.repo_full_name(mock_webhook)}.', mock_webhook
         )
 
 
 @mock.patch('harvey.utils.Utils.success')
-@mock.patch('harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='pull'), MOCK_OUTPUT, MOCK_TIME])  # noqa
+@mock.patch(
+    'harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='pull'), MOCK_OUTPUT, MOCK_TIME]
+)
 def test_start_pipeline_pull(mock_initialize_pipeline, mock_utils_success, mock_webhook):
     _ = Pipeline.start_pipeline(mock_webhook, False)
 
@@ -59,7 +60,9 @@ def test_start_pipeline_pull(mock_initialize_pipeline, mock_utils_success, mock_
 
 @mock.patch('harvey.utils.Utils.success')
 @mock.patch('harvey.pipelines.Pipeline.test')
-@mock.patch('harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='test'), MOCK_OUTPUT, MOCK_TIME])  # noqa
+@mock.patch(
+    'harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='test'), MOCK_OUTPUT, MOCK_TIME]
+)
 def test_start_pipeline_test(mock_initialize_pipeline, mock_test_pipeline, mock_utils_success, mock_webhook):
     _ = Pipeline.start_pipeline(mock_webhook, False)
 
@@ -70,7 +73,10 @@ def test_start_pipeline_test(mock_initialize_pipeline, mock_test_pipeline, mock_
 
 @mock.patch('harvey.utils.Utils.success')
 @mock.patch('harvey.pipelines.Pipeline.deploy', return_value=['mock-build', 'mock-deploy', 'mock-healthcheck'])
-@mock.patch('harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='deploy'), MOCK_OUTPUT, MOCK_TIME])  # noqa
+@mock.patch(
+    'harvey.pipelines.Pipeline.initialize_pipeline',
+    return_value=[mock_config(pipeline='deploy'), MOCK_OUTPUT, MOCK_TIME],
+)
 def test_start_pipeline_deploy(mock_initialize_pipeline, mock_deploy_pipeline, mock_utils_success, mock_webhook):
     _ = Pipeline.start_pipeline(mock_webhook, False)
 
@@ -82,8 +88,12 @@ def test_start_pipeline_deploy(mock_initialize_pipeline, mock_deploy_pipeline, m
 @mock.patch('harvey.utils.Utils.success')
 @mock.patch('harvey.pipelines.Pipeline.test')
 @mock.patch('harvey.pipelines.Pipeline.deploy', return_value=['mock-build', 'mock-deploy', 'mock-healthcheck'])
-@mock.patch('harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='full'), MOCK_OUTPUT, MOCK_TIME])  # noqa
-def test_start_pipeline_full(mock_initialize_pipeline, mock_deploy_pipeline, mock_test_pipeline, mock_utils_success, mock_webhook):  # noqa
+@mock.patch(
+    'harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='full'), MOCK_OUTPUT, MOCK_TIME]
+)
+def test_start_pipeline_full(
+    mock_initialize_pipeline, mock_deploy_pipeline, mock_test_pipeline, mock_utils_success, mock_webhook
+):
     _ = Pipeline.start_pipeline(mock_webhook, False)
 
     mock_initialize_pipeline.assert_called_once_with(mock_webhook)
@@ -93,7 +103,10 @@ def test_start_pipeline_full(mock_initialize_pipeline, mock_deploy_pipeline, moc
 
 
 @mock.patch('harvey.utils.Utils.kill')
-@mock.patch('harvey.pipelines.Pipeline.initialize_pipeline', return_value=[mock_config(pipeline='bad_name'), MOCK_OUTPUT, MOCK_TIME])  # noqa
+@mock.patch(
+    'harvey.pipelines.Pipeline.initialize_pipeline',
+    return_value=[mock_config(pipeline='bad_name'), MOCK_OUTPUT, MOCK_TIME],
+)
 def test_start_pipeline_bad_pipeline_name(mock_initialize_pipeline, mock_utils_kill, mock_webhook):
     _ = Pipeline.start_pipeline(mock_webhook, False)
 
@@ -119,20 +132,24 @@ def test_test_pipeline_error(mock_test_stage, mock_utils_kill, mock_webhook):
     mock_utils_kill.assert_called_once()
 
 
+@mock.patch('harvey.stages.DeployStage.run_container_healthcheck', return_value=True)
 @mock.patch('harvey.stages.DeployComposeStage.run')
-def test_deploy_pipeline_compose_success(mock_deploy_stage, mock_webhook):
-    _, _, _ = Pipeline.deploy(mock_config('deploy'), mock_webhook, MOCK_OUTPUT, MOCK_TIME, True)
+def test_deploy_pipeline_compose_success(mock_deploy_stage, mock_run_container_healthcheck, mock_webhook):
+    _, _, healthcheck = Pipeline.deploy(mock_config('deploy'), mock_webhook, MOCK_OUTPUT, MOCK_TIME, True)
 
+    assert healthcheck is True
     mock_deploy_stage.assert_called_once_with(mock_config('deploy'), mock_webhook, MOCK_OUTPUT)
 
 
-@mock.patch('harvey.stages.DeployStage.run_container_healthcheck')
+@mock.patch('harvey.stages.DeployStage.run_container_healthcheck', return_value=True)
 @mock.patch('harvey.stages.BuildStage.run')
 @mock.patch('harvey.stages.DeployStage.run')
-def test_deploy_pipeline_no_compose_success(mock_deploy_stage, mock_build_stage,
-                                            mock_run_container_healthcheck, mock_webhook):
-    _, _, _ = Pipeline.deploy(mock_config('deploy'), mock_webhook, MOCK_OUTPUT, MOCK_TIME, False)
+def test_deploy_pipeline_no_compose_success(
+    mock_deploy_stage, mock_build_stage, mock_run_container_healthcheck, mock_webhook
+):
+    _, _, healthcheck = Pipeline.deploy(mock_config('deploy'), mock_webhook, MOCK_OUTPUT, MOCK_TIME, False)
 
+    assert healthcheck is True
     mock_build_stage.assert_called_once_with(mock_config('deploy'), mock_webhook, MOCK_OUTPUT)
     mock_deploy_stage.assert_called_once_with(mock_webhook, MOCK_OUTPUT)
     mock_run_container_healthcheck.assert_called_once_with(mock_webhook)
@@ -142,8 +159,9 @@ def test_deploy_pipeline_no_compose_success(mock_deploy_stage, mock_build_stage,
 @mock.patch('harvey.stages.DeployStage.run_container_healthcheck', return_value=False)
 @mock.patch('harvey.stages.BuildStage.run')
 @mock.patch('harvey.stages.DeployStage.run')
-def test_deploy_pipeline_no_compose_failed_healthcheck(mock_deploy_stage, mock_build_stage,
-                                                       mock_run_container_healthcheck, mock_utils_kill, mock_webhook):
+def test_deploy_pipeline_no_compose_failed_healthcheck(
+    mock_deploy_stage, mock_build_stage, mock_run_container_healthcheck, mock_utils_kill, mock_webhook
+):
     _, _, healthcheck = Pipeline.deploy(mock_config('deploy'), mock_webhook, MOCK_OUTPUT, MOCK_TIME, False)
 
     assert healthcheck is False